# Go: fix data race on the coreClient with `sync.Mutex` and a channel map