--- a/go/integTest/shared_commands_test.go
+++ b/go/integTest/shared_commands_test.go
@@ -4595,7 +4595,8 @@ func (suite *GlideTestSuite) TestXRead() {
 		// but we wrap the test with timeout to avoid test failing or stuck forever
 		finished := make(chan bool)
 		go func() {
-			testClient.XReadWithOptions(map[string]string{key1: "0-1"}, *options.NewXReadOptions().SetBlock(0))
+			_, err := testClient.XReadWithOptions(map[string]string{key1: "0-1"}, *options.NewXReadOptions().SetBlock(0))
+			assert.IsType(suite.T(), &errors.ClosingError{}, err)
 			finished <- true
 		}()
 		select {
@@ -4604,6 +4605,7 @@ func (suite *GlideTestSuite) TestXRead() {
 		case <-time.After(3 * time.Second):
 		}
 		testClient.Close()
+		<-finished
 	})
 }