@rueian 2025-02-25
So I think this PR is ready. I did a simple benchmark:

```go
package glide

import (
	"github.com/valkey-io/valkey-glide/go/api"
	"testing"
)

func BenchmarkGlide(b *testing.B) {
	config := api.NewGlideClientConfiguration().WithAddress(&api.NodeAddress{Host: "127.0.0.1", Port: 6379})
	client, err := api.NewGlideClient(config)
	if err != nil {
		b.Fatal()
	}
	defer client.Close()

	b.ResetTimer()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			_, _ = client.Get("key")
		}
	})
	b.StopTimer()
}
```

Run the benchmark with the `go test -bench="Benchmark*" -benchmem -benchtime=100000x -count=10 . | tee 2_with_mutex.txt` command and using the `benchstat`:

```sh
▶ benchstat *
goos: darwin
goarch: arm64
pkg: benchmark/glide
cpu: Apple M1 Pro
         │ 1_without_mutex.txt │          2_with_mutex.txt          │
         │       sec/op        │   sec/op     vs base               │
Glide-10           8.190µ ± 1%   8.371µ ± 2%  +2.22% (p=0.001 n=10)

         │ 1_without_mutex.txt │          2_with_mutex.txt          │
         │        B/op         │    B/op     vs base                │
Glide-10            136.0 ± 0%   160.0 ± 1%  +17.65% (p=0.000 n=10)

         │ 1_without_mutex.txt │          2_with_mutex.txt          │
         │      allocs/op      │ allocs/op   vs base                │
Glide-10            4.000 ± 0%   5.000 ± 0%  +25.00% (p=0.000 n=10)
```

It shown that this PR can make 2% performance degradation on through put.